==> codes/2. DeNovo Negative Sampling.ipynb <==
# 2. DeNovo Negative Dataset Sampling for HP-PPI

Dissimilarity-based negative sampling for non-interacting protein pairs

Reference:

>Eid, F. E., Elhefnawi, M., & Heath, L. S. (2016). DeNovo: Virus-host sequence-based protein-protein interaction prediction. *Bioinformatics*, **32**(8), 1144–1150. https://doi.org/10.1093/bioinformatics/btv737

import os

import pandas as pd
import numpy as np
import itertools
import random

from time import time
from tqdm import tqdm
from collections import Counter
from multiprocessing import Pool

from Bio import SeqIO
from Bio.Align import PairwiseAligner
from Bio.SubsMat.MatrixInfo import blosum30

# Print status of datasets
def print_status(df):
    
    # For each pathogen
    for pathogen in sorted(set(df.Pathogen)):
        df_patho = df[df.Pathogen == pathogen]
        i = len(df_patho)
        p = len(set(df_patho.Pathogen_Uniprot_ID))
        h = len(set(df_patho.Human_Uniprot_ID))
        print('%s:\n%i non-interacting pairs involving %i pathogen proteins and %i human proteins\n' % (pathogen, i, p, h))
    
    # Total
    i = len(df)
    p = len(set(df.Pathogen_Uniprot_ID))
    h = len(set(df.Human_Uniprot_ID))
    print('TOTAL:\n%i non-interacting pairs involving %i pathogen proteins and %i human proteins\n' % (i, p, h))

# Set up directories
parent_dir = os.path.dirname(os.getcwd())

dir_in = os.path.join(parent_dir, 'raw_data', 'training')
dir_out = os.path.join(parent_dir, 'processed_data')

## Set up the sequences of pathogen proteins

# Read positive dataset
f_in = os.path.join(dir_out, 'positive_pairs.tsv')
df_pos = pd.read_csv(f_in, sep='\t')

# Get pathogen protein IDs for alignment
prot_IDs = list(set(df_pos.Pathogen_Uniprot_ID))
prot_IDs.sort(reverse=True)

print('Obtained %i pathogen proteins for alignment' % len(prot_IDs))

# Extract pathogen protein sequences
f_in = os.path.join(dir_in, 'uniprot_sequences.fasta')
seqdict = {} # store sequences in a dict

for record in SeqIO.parse(f_in, 'fasta'):
    prot_id = record.id.split('|')[1]
    
    # Check if record is in the pathogen IDs
    if prot_id in prot_IDs:
        seqdict[prot_id] = record.seq

print('Obtained %i pathogen protein sequences' % len(seqdict))

## All-vs-all global alignment of pathogen proteins

May skip to the next section (`Negative Sampling`) if alignment scores file already exists

# Set up aligner
aligner = PairwiseAligner()

# Gap scoring: linear, 8 (as in MATLAB, which is the platform used by Eid et al., 2016)
# Matrix: BLOSUM30 to capture distant relationships

aligner.gap_score = -10
aligner.substitution_matrix = blosum30
print('Alignment algorithm: %s' % aligner.algorithm)

# Generate pairs of proteins to be aligned
pairs = [pair for pair in itertools.combinations_with_replacement(prot_IDs, 2)]
print('Generated %i pairs for alignment' % len(pairs))

# Initialize zeros DataFrame for storing alignment scores
scores = pd.DataFrame(0., index=prot_IDs, columns=prot_IDs)

scores.shape
scores.head()

# Function to get the alignment score
def get_alignment_score(pair):
    score = aligner.score(*map(seqdict.get, pair))
    
    # Return both proteins along with the alignment score
    return (*pair, score)

# Perform alignments with parallelization
with Pool(6) as p:
    t0 = time()
    
    results = p.map(get_alignment_score, pairs)
    
    t = time() - t0
    print('Alignment duration: %.4f minutes' % (t/60))

# Store alignment scores
for prot1, prot2, score in tqdm(results):
    scores.at[prot1, prot2] = score
    scores.at[prot2, prot1] = score

scores.head()

# Save the scores DataFrame
f_out = os.path.join(dir_out, 'blosum30_global_scores.tsv')
scores.to_csv(f_out, sep='\t', index=prot_IDs)

## Negative sampling

Select non-interacting pairs based on dissimilarity between pathogen protein interactors

# Load the scores DataFrame
f_in = os.path.join(dir_out, 'blosum30_global_scores.tsv')
scores = pd.read_csv(f_in, sep='\t', index_col=0)

# Designate outlier scores as NaN (protein with very low alignment score compared to others)
# Keep removing outliers as long as it has minimum scores across x% of total pathogen proteins
x = 0.99

while True:
    row_mins = scores.idxmin(axis=1)
    outlier_id = row_mins.mode()[0] # a bug: axis=0 actually considers columns instead of rows
    
    n_outlier = Counter(row_mins)[outlier_id]
    if n_outlier > int(x * len(prot_IDs)):
        
        # Remove all entries in the outlier column except for its own row
        scores.loc[~(scores.index == outlier_id), outlier_id] = np.nan
        print('Removed outlier protein: %s' % outlier_id)
    
    else:
        print('No more outliers found.')
        break

# Calculate dissimilarity distances: complement of min-max scaled scores
norm_scores = (scores - scores.min(axis=1)) / (scores.max(axis=1) - scores.min(axis=1))
distance = 1 - norm_scores

distance.head()

# Prepare positive pairs and tag species to each protein
positive_pairs = df_pos.drop('Pathogen', axis=1).values.tolist()

d = df_pos[['Pathogen', 'Pathogen_Uniprot_ID']].values
species_tag = {protein: species for species, protein in d}

# DeNovo negative sampling
# T: dissimilarity threshold (only consider proteins with distance < T)
# n: maximum number of proteins to be selected (relative to its occurence in positive dataset)
T = 0.8
n = 20

negative_pairs = []
for pathogen_prot in tqdm(prot_IDs):
    # Set up: pathogen protein count and tagging
    c = df_pos.Pathogen_Uniprot_ID.tolist().count(pathogen_prot)
    tag = species_tag[pathogen_prot]
    
    # Get a list of pathogen interactors included for sampling
    cond = distance.loc[pathogen_prot] > 0.8
    patho_inc = cond[cond].index
    
    # Get list of human proteins that interacts with patho_inc
    df = df_pos[df_pos.Pathogen_Uniprot_ID.isin(patho_inc)]
    human_inc = set(df.Human_Uniprot_ID)
    
    # Pair up the current pathogen protein with human_inc
    temp_pairs = []
    for pair in itertools.product([pathogen_prot], human_inc):
        if list(pair) not in positive_pairs:
            temp_pairs.append((tag, *pair))
    
    # Sample pairs for negative dataset
    N = n * c
    if len(temp_pairs) > N:
        negative_pairs += random.sample(temp_pairs, N)
    else:
        negative_pairs += temp_pairs

# Examine negative pairs DataFrame
df_neg = pd.DataFrame(negative_pairs, columns=['Pathogen', 'Pathogen_Uniprot_ID', 'Human_Uniprot_ID'])

print_status(df_neg)

# Save negative pairs
f_out = os.path.join(dir_out, 'negative_pairs.tsv')
df_neg.to_csv(f_out, sep='\t', index=False)
df_neg.head()

<hr></hr>
