==> codes/3. Feature Extraction.ipynb <==
# 3. Extraction of Domain Features for Training Dataset

Generate feature vectors of domain occurences in each protein pair

import os
import joblib

import pandas as pd
import numpy as np
from scipy import sparse
from tqdm import tqdm
from functools import partial

from features import domain_features

# Print status of datasets
def print_status(df):
    
    # For each pathogen
    for pathogen in sorted(set(df.Pathogen)):
        df_patho = df[df.Pathogen == pathogen]
        i = len(df_patho)
        p = len(set(df_patho.Pathogen_Uniprot_ID))
        h = len(set(df_patho.Human_Uniprot_ID))
        print('%s:\n%i pairs involving %i pathogen proteins and %i human proteins\n' % (pathogen, i, p, h))
    
    # Total
    i = len(df)
    p = len(set(df.Pathogen_Uniprot_ID))
    h = len(set(df.Human_Uniprot_ID))
    print('TOTAL:\n%i pairs involving %i pathogen proteins and %i human proteins\n' % (i, p, h))

# Set up directories
parent_dir = os.path.dirname(os.getcwd())

dir_in = os.path.join(parent_dir, 'processed_data')
dir_out = os.path.join(parent_dir, 'processed_data', 'features')

# Load positive dataset
f_in = os.path.join(dir_in, 'positive_pairs.tsv')
df_pos = pd.read_csv(f_in, sep='\t')

print('POSITIVE DATASET\n'.center(70))
print_status(df_pos)

# Load negative dataset
f_in = os.path.join(dir_in, 'negative_pairs.tsv')
df_neg = pd.read_csv(f_in, sep='\t')

print('NEGATIVE DATASET\n'.center(70))
print_status(df_neg)

## Transform protein pairs into feature vectors

Features:
- domain profiles

# Load Pfam data
pfam_dict, pfam_set = joblib.load('pfam.pkl')

# Get features of each protein as dict
feature_function = partial(domain_features,
                           domain_dict=pfam_dict,
                           domain_set=pfam_set)

feat_dict = {prot: feature_function(prot) for prot in tqdm(pfam_dict.keys())}

# Feature extraction: domain occurences

def pairs_to_features(protein_pairs, label):
    '''Transform protein pairs into a vector of features and add labels.
    Labels:
    - 1 for positive dataset
    - 0 for negative dataset
    '''
    
    # Initialize an empty list for features
    X = []
    for pair in tqdm(protein_pairs):
        features = sum(map(feat_dict.get, pair))
        X.append(features)
    
    # Save features as a sparse matrix
    X = sparse.vstack(X)
    
    # Generate labels
    y = np.array([label] * X.shape[0])
    
    return X, y

# Iterate feature extraction by dataset and pathogen

d = {'pos': (df_pos, 1), 'neg': (df_neg, 0)} # df and label
for pathogen in set(df_pos.Pathogen):
    
    # Store positive and negative feature arrays in a list
    X = []
    y = []
    
    # Iterate through positive & negative for pathogen
    for data in d.keys():
        df, label = d[data]
        
        # Select only the current pathogen from datasets
        df_current = df[df.Pathogen == pathogen]
        pairs = df_current.drop('Pathogen', axis=1).values

        # Extract features
        X_, y_ = pairs_to_features(pairs, label)
        X.append(X_)
        y.append(y_)
    
    # Combine positive and negative features and labels
    X = sparse.vstack(X)
    y = np.concatenate(y)
    
    # Dump data as pickle
    ## Slice pathogen species name for file name
    genus, species = pathogen.split()[:2]
    name = genus[:3] + species[:2]

    f_out = os.path.join(dir_out, '%s_features.pkl' % name)
    joblib.dump((X, y), f_out)

<hr></hr>
