==> codes/5. Prediction and Analysis of Protein Interaction.ipynb <==
# Prediction of Protein Interaction Network

Machine learning classification task for predicting whether a protein pair interacts or not

Protein pairs are generated from a test pathogen against human proteins contained in training set

import os
import joblib
import itertools
from functools import partial
from multiprocessing import Pool
from tqdm import tqdm
from time import time

import pandas as pd
import numpy as np
from scipy import sparse

from Bio import SearchIO

from features import domain_features

# Set up directories
parent_dir = os.path.dirname(os.getcwd())

dir_in = os.path.join(parent_dir, 'raw_data', 'prediction')
dir_out = os.path.join(parent_dir, 'analyses')

## Obtain candidate protein domains

Protein sequence source: https://www.uniprot.org

Pathogen: *Streptococcus pneumoniae* strain D39 (**STRP2**; Taxonomy ID 373153)
>Keyword filters
- `Organism [OS]: Streptococcus pneumoniae serotype 2 (strain D39 / NCTC 7466) [373153]`
- `Sequence length: From 50`
- `Gene Ontology [GO]: extracellular region [5576] (20), pathogenesis [9405] (3)`
- `All: choline (9), sialidase (2), amidase (4), virulence (4), surface protein (2), adhesion (3)`
- Total: 42 proteins

Host: *Homo sapiens* (Taxonomy ID 9606)
>Keyword filters
- `Organism [OS]: Homo sapiens (Human) [9606]`
- `Proteome ID: up000005640`
- `Sequence length: From 50`
- `All: extracellular`
- `Gene Ontology [GO]: immunoglobulin receptor activity [19763] (6), regulation of complement activation [30449] (120), phagocytosis [6909] (204), toll-like receptor signalling pathway [2224] (54), plasminogen activation [31639] (15), defense response to Gram-positive bacterium [50830] (87)`
- Total: 395 proteins


Extraction of Pfam domains: `hmmscan`
- `hmmscan --tblout STRP2_pfam_hits --acc --noali -E 0.00001 --domE 0.00001 --cpu 7 ~/hmmer-3.2.1/pfam/Pfam-A.hmm STRP2_sequences.fasta`

# Parse hmmscan result of STRP2
pfam_set = joblib.load('pfam.pkl')[1] # Pfam domains available in the training dataset
prot_lists = {}
pfam_dict = {} # generate a new dict for the candidate proteins

for organism in ['STRP2', 'HUMAN']:
    f_in = os.path.join(dir_in, '%s_pfam_hits' % organism)
    proteins = [] # store Uniprot accessions of the current organism
    
    for query in SearchIO.parse(f_in, 'hmmer3-tab'):
        uniprot_id = query.id.split('|')[2] # use protein name to ease network analysis
        domain_counts = {}

        # Read each domain hits in query
        for hit in query.hits:
            pfam_acc = hit.accession.split('.')[0] # Pfam accession of domain

            # Select only domains that exists in the training set
            if pfam_acc in pfam_set: 
                domain_counts[pfam_acc] = hit.domain_reported_num

                # Add the pathogen protein to an existing domains dict
                pfam_dict[uniprot_id] = domain_counts
        
                if uniprot_id not in proteins:
                    proteins.append(uniprot_id)
    
    # Store protein list of the current organism in a dict
    prot_lists[organism] = proteins
    
    # Print statistics
    print('Selected %i %s proteins for HP-PPI prediction' % (len(proteins), organism))
    
# Save new Pfam dict
_ = joblib.dump(pfam_dict, 'prediction_pfam.pkl')

## Generate candidate protein pairs and extract domain features

# Generate protein pairs
pairs = [pair for pair in itertools.product(prot_lists['STRP2'], prot_lists['HUMAN'])]
print('Generated %i protein pairs for prediction' % len(pairs))

# Set up feature extraction function
feature_function = partial(domain_features,
                           domain_dict=pfam_dict,
                           domain_set=pfam_set)

# Get features of each protein as dict to speed up
# extraction from pairs
all_prots = prot_lists['STRP2'] + prot_lists['HUMAN']
feat_dict = {prot: feature_function(prot) for prot in tqdm(all_prots)}

## HP-PPI Prediction

# Load pre-trained classifier
pipe = joblib.load('best_model.pkl')
pipe['clf'].get_params()

# Prediction task
X = []
for pair in tqdm(pairs):
    features = sum(map(feat_dict.get, pair))
    X.append(features)

X = sparse.vstack(X)

# Predict labels
predictions = pipe.predict(X)

# Examine prediction result
n_ppi = predictions.sum() # number of interactions reported
print('Predicted %i interactions (%.2f%% of candidate pairs)' % (n_ppi, (n_ppi/len(pairs) * 100)))

# Obtain interacting pairs
interactions = np.array(pairs)[predictions.astype(bool)]

# Save interactions as DataFrame
f_out = os.path.join(dir_out, 'predicted_interactions.tsv')

df = pd.DataFrame(interactions, columns=['Pathogen_Protein', 'Human_Protein'])
df.to_csv(f_out, sep='\t', index=False)
df.head()

<hr></hr>
