==> codes/6. Network Analysis.ipynb <==
# 6. Analysis of the Predicted HP-PPI Network

Network analysis on the human-*Streptococcus pneumoniae* strain D39 protein-protein interaction network with NetworkX module

import os
import joblib

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import networkx as nx
from networkx.algorithms import bipartite
from pyvis.network import Network

# Set up directories
parent_dir = os.path.dirname(os.getcwd())
dir_in = dir_out = os.path.join(parent_dir, 'analyses')

# Load predicted network as DataFrame
f_in = os.path.join(dir_in, 'predicted_interactions.tsv')
df = pd.read_csv(f_in, sep='\t')

# Get sets of unique proteins from each organism
pathogen_proteins = list(set(df.Pathogen_Protein))
human_proteins = list(set(df.Human_Protein))

print('Loaded %i interactions involving %i STRP2 proteins and %i human proteins\n'\
      % (len(df), len(pathogen_proteins), len(human_proteins)))
df.head()

# Save list of proteins
for organism in ['Pathogen', 'Human']:
    f_out = os.path.join(dir_out, 'protein_lists', '%s_protein_list' % organism)
    protein_list = set(df['%s_Protein' % organism])
    with open(f_out, 'w') as f:
        f.write('\n'.join(protein_list))

## Create graph of protein interactions

# Rename proteins by gene names
for organism in ['STRP2', 'HUMAN']:
    f_in = os.path.join(dir_in, 'protein_lists', '%s_uniprot_mapping.tab' % organism)
    
    gene_mapping = pd.read_csv(f_in, sep='\t')[['Entry', 'Gene names']].values
    map_dict = {uniprot: gene.split(' ')[0] for uniprot, gene in gene_mapping}
    df.replace(map_dict, inplace=True)

df.head()

# Initialize a bipartite graph 
pathogen_proteins = list(set(df.Pathogen_Protein))
human_proteins = list(set(df.Human_Protein))
B = nx.Graph()

# Add proteins as nodes with a bipartite attribute
B.add_nodes_from(pathogen_proteins, bipartite=0)
B.add_nodes_from(human_proteins, bipartite=1)

# Add interactions as edges between nodes
B.add_edges_from(df.values)

# Create DataFrame to store properties of pathogen proteins
f_in = os.path.join(dir_in, 'protein_lists', 'STRP2_uniprot_mapping.tab')

df = pd.read_csv(f_in, sep='\t')[['Gene names', 'Protein names']]
df['Gene names'] = df['Gene names'].apply(lambda x: x.split(' ')[0])

_ = df.set_index('Gene names', inplace=True)
df.head()

# Analyze topological properties
properties = {
    'Degree_centrality': bipartite.degree_centrality,
    'Betweenness_centrality': bipartite.betweenness_centrality
}

for prop, func in properties.items():
    props = func(B, nodes=pathogen_proteins)
    df_prop = pd.DataFrame.from_dict(props, orient='index', columns=[prop])
    
    # Obtain only pathogen proteins
    df_prop = df_prop[df_prop.index.isin(pathogen_proteins)]
    
    # Concatenate to the empty DataFrame
    df = pd.concat([df, df_prop], axis=1, sort=False)

# Examine top-scoring properties
df.index.name = 'Gene_name'
df.sort_values(by='Degree_centrality', ascending=False, inplace=True)
df

# Save centralities
f_out = os.path.join(dir_out, 'network_centrality.tsv')
df.to_csv(f_out, sep='\t')

# Get enriched human proteins
f_in = os.path.join(dir_in, 'KEGG_pathway_enrichment.txt')
df_kegg = pd.read_csv(f_in, sep='\t')
df_kegg.Term = df_kegg.Term.apply(lambda x: x.split(':')[1])

# Set Term column as index
df_kegg.set_index('Term', inplace=True)

# Obtain only relevant pathways
pathways = ['Complement and coagulation cascades',
            'Phagosome',
            'Fc gamma R-mediated phagocytosis',
            'Bacterial invasion of epithelial cells']

# Create subgraph for each pathway
colors = ['red', 'orange', 'gold', 'khaki']

for i, term in enumerate(pathways):
    human_proteins = df_kegg.loc[term]['Genes'].split(', ')
    subgraph = B.subgraph(pathogen_proteins + human_proteins)

    # Visualize the network
    net = Network()
    net.inherit_edge_colors_from(False)
    
    # Color human & pathogen nodes differently
    m = len(human_proteins)
    n = len(pathogen_proteins)
    net.add_nodes(pathogen_proteins, color=['blue']*n)
    net.add_nodes(human_proteins, color=[colors[i]]*m)
    net.add_edges(subgraph.edges())
    
    # Miscellaneous graph settings
    net.force_atlas_2based(gravity=-200, spring_length=150, overlap=0.2)
    
    net.show_buttons()
    net.show('subgraph_%i.html' % (i+1))

<hr></hr>
